Update of the motion correction : the brick now support zipped files. A new phase in the tests is now reached : the estimated parameters based on minc tracc are very close of the ones estimated through SPM5, so the code can be considered pretty much ready to be used.
Regarding the import of the realignement procedure based on SPM_REALIGN (the NIAK_BRICK_REALIGN), this project is put on 
hold for an undetermined period of time, as the MNC2NII conversion does not really seem to work (flipping problems).

The style of the help of the function has also been revisited.



git-svn-id: http://niak.googlecode.com/svn/trunk@475 3762db0e-9444-0410-bee5-d3813040ea1c