a couple of minor modifs.
The main thing of this commit is NIAK_BRICK_MOTION_CORRECTION_DEV which should be operational. I've put 
some efforts to speed it up. Rigid-body motion parameters estimation should take a couple minutes for one 
run on a single machine. It is still much slower than SPM (roughly 2 to 3 times slower I 
would say) but the difference is not dramatic because that is in any case pretty fast. Anyway it is very 
hard to speed up the algorithm beyond that point because most of the computation time is internal to 
MINCTRACC.



git-svn-id: http://niak.googlecode.com/svn/trunk@771 3762db0e-9444-0410-bee5-d3813040ea1c