Big commit : essentially the implementation of a bunch of stuff related to brain extraction in T1 images has been revised to work with Octave.
That was a much bigger job than expected, because the morphomat tools in Octave are not quite there. Finally I decided to rely 
heavily on MINCMORPH, so the algorithm is now much slower. But it seems to work both in Matlab and Octave. I still have a couple of 
tests to do though.



git-svn-id: http://niak.googlecode.com/svn/trunk@992 3762db0e-9444-0410-bee5-d3813040ea1c