Work-in-progress: slight modification of the design of the connectome pipeline. The options are going to be much simpler to enter.
There are going to be flags to turn on/off certain categories of measures (p2p / rmap / local_prop / global_prop). 

For all measures but global_prop (aka global network properties), it will be necessary to enter a csv file with either some labels of parcels or some coordinates in stereotaxic space (which indirectly define a parcel). All possible p2p/rmap/local_prop associated with these seeds will then be generated. It will thus be pretty simple to write a script for the pipeline, while now it is a mess of labels / parcel indices / measure names. 




git-svn-id: http://niak.googlecode.com/svn/trunk@1764 3762db0e-9444-0410-bee5-d3813040ea1c